+++ lib/index.js
@@ -1,14 +1,10 @@
 require('./promise')
 
-const cluster = require('cluster')
-const os = require('os')
-const http = require('http')
-const promClient = require('prom-client')
-
 const EventEmitter = require('./event_emitter')
 const Errors = require('./errors/errors')
 const Container = require('./di/container')
 const Profiler = require('./profiler')
+const { init: initClustering } = require('./clustering')
 let container = new Container()
 
 const Plugin = require('./plugin')
@@ -19,8 +15,6 @@ global.Q = {
   profiler: new Profiler(),
   container
 }
-
-const aggregatorRegistry = new promClient.AggregatorRegistry()
 
 container.registerSingleton('app', require('./app'))
 container.registerSingleton('config', require('./config/config'))
@@ -34,32 +28,6 @@ process.on('unhandledRejection', function(reason, promise) {
   }
 })
 
-function metricsServerHandler(req, res) {
-  if (req.method === 'GET' && req.url === '/metrics') {
-    res.writeHead(200, { 'Content-Type': 'text/plain' })
-    res.end(promClient.register.metrics())
-  } else {
-    res.writeHead(404).end()
-  }
-}
-
-function aggregatedMetricsServerHandler(req, res) {
-  if (req.method === 'GET' && req.url === '/metrics') {
-    aggregatorRegistry.clusterMetrics((err, metrics) => {
-      if (err) {
-        Q.log.warn(err)
-        res.writeHead(500, 'Content-Type', 'text/plain')
-        res.end(err.message)
-      } else {
-        res.writeHead(200, 'Content-Type', aggregatorRegistry.contentType)
-        res.end(metrics)
-      }
-    })
-  } else {
-    res.writeHead(404).end()
-  }
-}
-
 async function boot(opts) {
   opts = opts || {}
 
@@ -71,33 +39,9 @@ async function boot(opts) {
   Q.config.loadFrom(Q.plugins.map(_ => _.pluginDir))
 
   Q.log = await container.getAsync('log')
-  const cmd = Q.app.getAppCommand()
 
-  // start workers
-  const clustering = (
-    Q.config.get('quadro.clustering', false) &&
-    ['test', 'repl', 'task'].indexOf(cmd) < 0
-  )
-
-  if (cluster.isMaster && clustering) {
-    for (let i = 0; i < os.cpus().length; ++i) cluster.fork()
-    cluster.on('exit', (worker, code, signal) => {
-      Q.log.info(`Worker ${worker.process.pid} exited with code ${code} from signal ${signal}.`)
-      process.exit(code)
-    })
-  }
-
-  // start prometheus server
-  if (clustering && cluster.isMaster) {
-    http.createServer(aggregatedMetricsServerHandler)
-      .listen(Q.config.get('prometheus.aggregatorPort', 9230))
-  } else if (!clustering) {
-    http.createServer(metricsServerHandler)
-      .listen(Q.config.get('prometheus.port', 9230))
-  } // else, there is clustering, and this is not the master, so do nothing
-
-  // prevent master from doing work
-  if (clustering && cluster.isMaster) return
+  // halt init if this is the master and we are clustering
+  if (initClustering()) return
 
   await container.run(require('./services_loader'))
   await container.run(require('./models_loader'))
@@ -109,6 +53,7 @@ async function boot(opts) {
 
   await Q.app.emit('loaded', Q.app)
 
+  const cmd = Q.app.getAppCommand()
   if (cmd === 'test') return runTests()
   if (cmd === 'repl') return runREPL()
   if (cmd === 'task') return runTask()
